Created ErrorHandler.ts - ErrorHandler handles catch(err) when making a HTTP request - ErrorHandler returns the error message to display to the user - Refactor code in AuthUser.ts & meal-service.ts to utlize the ErrorHandler.ts
