+++ src/services/meal-service.ts
@@ -8,16 +8,13 @@ import axios, { AxiosResponse } from 'axios';
 import type { IMeal } from '../types/meal-types';
 import type { AuthUser } from '../types/users-types';
 
-interface IErrorResponse {
-    // Interface for response data from a HTTP request when we get an error
-    message: string;
-}
-
 interface GetUserMealData {
+    // Interface for the data recieved when requesting a user's meals
     meals: IMeal[]
 }
 
 interface CreateMealData{
+    // Data recieved when sending a new meal to the backend
     meal: IMeal
 }
 
@@ -44,34 +41,6 @@ async function createMeal(user:AuthUser, meal: IMeal) {
     } catch(err){
         return null;
     }
-}
-
-
-const errorHandler = (error: any): string => {
-    /**
-     * Function handles the output error message given an error from a catch block
-     * @param error is an error from a catch block when making a async-await request
-     * @return String that will be the output error message to display
-    */
-
-    let errorMessage:string;
-         if (error.response){
-             // Request made and server responded, response.status has the HTTP error code we got back
-             const response:AxiosResponse = error.response;      // Save the response as AxiosResponse type
-             const data:IErrorResponse = response.data;          // Get the data from the response
-             errorMessage = data.message;                        // Get the error message from the data
-             return errorMessage;                                // Return the error to display in the app
-         }else if (error.request){
-             // The request was made but no response was received from the server
-            //  console.log(error.request);
-             errorMessage = "There was an error on our side, please try again later."
-             return errorMessage;
-         } else{
-             // Something happened in setting up the request that triggered an Error
-             console.log('Error', error.message);
-             errorMessage = "Some error occurred, please try again later."
-             return errorMessage;
-         }
 }
 
 export {
